Filtered transcript count data

This dataset contains transcript-level count data obtained with Salmon, imported using tximport, and annotated with Ensembl gene identifiers. Duplicate entries in gene biotypes were previously removed to ensure consistency in downstream analyses.

Usage

filtered_txi

Format

A list object with the following elements:

counts

Matrix of transcript counts.

abundance

Matrix of normalized transcript abundances.

length

Effective transcript lengths.

metadata

Data frame containing sample-level metadata.

Source

Processed internally from transcript quantifications generated with Salmon.

Details

This object is typically used for differential gene expression analysis with DESeq2, after importing transcript quantifications and mapping them to Ensembl identifiers.

Examples

data(filtered_txi)
names(filtered_txi)
#> [1] "abundance"           "counts"              "length"             
#> [4] "countsFromAbundance"
head(filtered_txi$counts)
#>          QR011_0   QR011_1  QR011_3 QR013_0   QR013_1   QR013_3 QR014_0
#> A1BG     179.000   134.000  401.000 163.000   221.687   240.000  96.000
#> A1BG-AS1  69.977    76.949  205.607  76.942   183.012   203.333  62.904
#> A1CF       0.000     0.000    0.000   0.000     0.000     0.000   0.000
#> A2M       16.000 27777.937 3700.000  29.584 46051.855 22768.000   9.000
#> A2M-AS1   11.000    14.063   12.000  24.416     9.145     0.000  15.000
#> A2ML1      0.000     0.000    0.000   0.000     0.000     0.000   0.000
#>            QR014_1   QR014_3 QR017_0   QR017_1   QR017_3 QR020_0   QR020_1
#> A1BG       112.892   121.239 185.000   278.591   137.888 204.743   280.851
#> A1BG-AS1   138.049   122.018  55.976   146.853   153.676  76.326   246.989
#> A1CF         0.000     0.000   0.000     0.000     0.000   0.000     0.000
#> A2M      41087.453 26001.863  28.000 21718.651 17429.000  59.959 41249.991
#> A2M-AS1      4.547     7.137  44.000     4.349    18.000  24.041     7.009
#> A2ML1        1.000     0.000   0.000     0.000     1.000   0.000     0.000
#>           QR020_3 QR021_0   QR021_1  QR021_3 QR022_0   QR022_1   QR022_3
#> A1BG      104.516  80.000   115.000  243.670 184.000    31.000   103.000
#> A1BG-AS1  147.253  49.772   159.942  225.720  22.322    21.694    50.307
#> A1CF        0.000   2.000     0.000    0.000   0.000     0.000     0.000
#> A2M      6317.077  29.000 16291.910 5344.043  73.000 38089.999 16589.194
#> A2M-AS1    19.923   7.000     5.090   31.957  32.000     4.001    42.806
#> A2ML1       0.000   0.000     0.000    0.000   0.000     0.000     0.000
#>          QR023_0   QR023_1   QR023_3 QR024_0  QR024_1   QR024_3 QR025_0
#> A1BG     357.000   190.794   107.000 136.000  122.751   143.000 138.000
#> A1BG-AS1 124.303   120.695   110.329  52.984  109.739    57.537  22.752
#> A1CF       0.000     0.000     0.000   0.000    0.000     0.000   0.000
#> A2M        5.000 12433.000 11389.913  22.596 8843.011 12490.831  23.000
#> A2M-AS1   35.000     5.000     9.087  24.404   24.989    18.168  17.000
#> A2ML1      0.000     0.000     0.000   0.000    0.000     0.000   0.000
#>            QR025_1  QR025_3 QR026_0   QR026_1  QR026_3 QR028_0   QR028_1
#> A1BG       148.675  156.805 113.000   109.215  230.771 442.000   188.722
#> A1BG-AS1    93.074  137.146  71.205   139.078  239.612 103.055   244.920
#> A1CF         0.000    0.000   0.000     0.000    0.000   0.000     0.000
#> A2M      58259.947 5720.999  14.000 36282.589 7888.874  50.000 37797.421
#> A2M-AS1     12.053   32.001   4.000    10.412   31.126  37.000    25.579
#> A2ML1        0.000    1.000   0.000     0.000    0.000   0.000     1.000
#>           QR028_3 QR060_0   QR060_1  QR060_3 QR061_0   QR061_1  QR061_3 QR062_0
#> A1BG       78.000 376.000   610.000  164.000 171.974   225.743   77.000 225.000
#> A1BG-AS1  137.076 131.025   430.112  216.977  64.499   240.434  172.862 116.364
#> A1CF        0.000   0.000     0.000    0.000   0.000     0.000    0.000   0.000
#> A2M      4456.345  31.000 12441.695 2596.596  16.000 21328.987 1048.994  76.739
#> A2M-AS1    21.654  26.000    34.304   47.404  18.000     9.013   16.006  29.261
#> A2ML1       0.000   0.000     0.000    0.000   0.000     1.000    0.000   0.000
#>            QR062_1  QR062_3 QR064_0   QR064_1  QR064_3 QR065_0  QR065_1 QR065_3
#> A1BG       132.505  202.878 123.000   195.913  146.000 187.000  117.000 423.987
#> A1BG-AS1   115.057  179.373  54.707   165.954  180.705  68.109  119.825 145.204
#> A1CF         0.000    0.000   0.000     0.000    0.000   0.000    1.000   2.000
#> A2M      29018.993 8332.050  18.000 29891.895 2297.000  10.000 8761.287 667.409
#> A2M-AS1     20.008   31.949  17.000    19.105   31.000  20.000   17.713  12.591
#> A2ML1        2.000    2.000   0.000     0.000    0.000   0.000    2.000   0.000
#>          QR077_0   QR077_1  QR077_3 QR080_0   QR080_1   QR080_3 QR093_0
#> A1BG      88.881   107.000  126.000 207.000   165.979   209.000  323.00
#> A1BG-AS1  44.679   106.320  133.388 113.115   127.906   180.334  115.06
#> A1CF       0.000     0.000    0.000   0.000     0.000     0.000    0.00
#> A2M        8.000 31486.894 5967.520  27.592 33747.000 10953.753   24.00
#> A2M-AS1    4.000     7.105   18.480  39.408    21.000    16.247   20.00
#> A2ML1      0.000     2.000    0.000   0.000     0.000     0.000    0.00
#>            QR093_1   QR093_3 QR094_0  QR094_1  QR094_3 QR095_0   QR095_1
#> A1BG       284.000   133.969 256.000  118.751  232.945 174.000    93.000
#> A1BG-AS1   215.357   138.505  83.129  131.000  162.490  83.989   101.517
#> A1CF         0.000     0.000   0.000    0.000    0.000   0.000     0.000
#> A2M      40348.881 11271.078  20.000 7487.000 7242.663  23.102 10449.960
#> A2M-AS1     37.119     7.922  33.000    6.000   13.338  46.898    17.041
#> A2ML1        0.000     3.000   0.000    0.000    0.000   0.000     0.000
#>           QR095_3 QR096_0   QR096_1  QR096_3 QR097_0   QR097_1  QR097_3 QR098_0
#> A1BG      135.980 154.857   101.897  169.954  94.899   141.000  145.000 184.891
#> A1BG-AS1   97.698  55.028    93.855  112.171  60.858   123.146  119.288  57.292
#> A1CF        0.000   0.000     0.000    0.000   0.000     0.000    0.000   0.000
#> A2M      1709.701  26.554 13213.712 2384.579  11.000 12632.828 5349.999  20.538
#> A2M-AS1    44.298  51.445    22.288   23.421  16.000     8.173   11.001  39.462
#> A2ML1       0.000   0.000     0.000    0.000   0.000     1.000    0.000   0.000
#>            QR098_1  QR098_3 QR100_0   QR100_1  QR100_3 QR102_0   QR102_1
#> A1BG       150.000  185.835 261.941   187.988  127.000 348.952   170.000
#> A1BG-AS1    91.023  163.192  79.100   148.057  135.507 103.671   154.111
#> A1CF         0.000    0.000   0.000     0.000    0.000   0.000     0.000
#> A2M      24509.000 1164.386  23.000 22239.989 8894.000  70.000 28735.310
#> A2M-AS1     11.000   24.614  46.000     6.011    6.000  53.000    10.689
#> A2ML1        0.000    0.000   0.000     1.000    0.000   0.000     1.000
#>            QR102_3 QR103_0   QR103_1  QR103_3 QR104_0   QR104_1   QR104_3
#> A1BG       270.406 153.000    94.860  146.906 209.000    52.000    80.000
#> A1BG-AS1   220.621  68.605   115.891  127.163  86.455    65.121    77.594
#> A1CF         0.000   0.000     0.000    0.000   0.000     0.000     0.000
#> A2M      13722.870  22.000 15330.628 6125.996  13.750 31269.962 21872.999
#> A2M-AS1     41.130  24.000    13.372   35.004   8.250     7.037     3.000
#> A2ML1        0.000   0.000     0.000    0.000   0.000     1.000     0.000
#>          QR105_0   QR105_1  QR105_3 QR106_0   QR106_1   QR106_3 QR107_0
#> A1BG      160.00   174.000  172.000 176.000   136.000   113.938 259.000
#> A1BG-AS1   67.63   138.282  101.998  97.486   101.134   119.771 137.703
#> A1CF        0.00     0.000    0.000   0.000     0.000     0.000   0.000
#> A2M         6.00 30761.132 6616.023  16.000 28952.000 17332.000  26.000
#> A2M-AS1    25.00    11.868   30.976  12.000     7.000     8.000  52.000
#> A2ML1       0.00     0.000    0.000   0.000     0.000     0.000   0.000
#>            QR107_1  QR107_3 QR108_0   QR108_1  QR108_3
#> A1BG       145.545  133.000 346.741   221.952  160.000
#> A1BG-AS1   193.229  215.026 129.217   139.239  205.754
#> A1CF         0.000    0.000   0.000     0.000    0.000
#> A2M      29487.916 8381.828  47.000 37502.065 7105.960
#> A2M-AS1     19.084   27.172  53.000    26.935   32.040
#> A2ML1        0.000    1.000   0.000     0.000    0.000